django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "AI-Powered Freelancing Platform Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "django-filter>=24.3",
    "django-redis>=5.4.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "requests>=2.32.0",
    "google-generativeai>=0.8.3",
    "pinecone-client>=5.0.1",
    "pgvector>=0.2.4",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "stripe>=11.1.0",
    "paypalrestsdk>=1.13.3",
    "python-jose>=3.3.0",
    "cryptography>=43.0.3",
    "pillow>=11.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.1",
    "djangorestframework-simplejwt>=5.3.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.2",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "neo4j>=5.15.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
]
